+++ src/partials/modal.html
@@ -1,18 +1,26 @@
 <div class="backdrop is-hidden" data-modal>
   <div class="container">
     <div class="modal-window">
-      <!-- <button type="button" class="close-modal-btn">
-        <svg class="close-btn-icon">
-          <use href="./images/sprite.svg#icon-close"></use>
-        </svg>
-      </button> -->
       <button type="button" class="close-modal-btn">
         <svg class="close-btn-icon">
           <use href="./images/sprite.svg#icon-close"></use>
         </svg>
       </button>
 
-      <div class="film-info-container"></div>
+      <div class="film-info-container">
+        <div class="img-wrap"></div>
+        <div class="film-info-wrap">
+          <div class="film-info-js"></div>
+          <div class="film-btn-wrap">
+            <button type="button" class="film-btn film-btn-wached">
+              add to Watched
+            </button>
+            <button type="button" class="film-btn film-btn-queue">
+              add to queue
+            </button>
+          </div>
+        </div>
+      </div>
     </div>
   </div>
 </div>
